refactor: RequestReader 객체 분리
